refactor frames to not use JS for handling element sizing
all sizing is now done in CSS. For that, the actual element that will be
overlayd by the frame, has been wrapped within a wrapper element, that
can be styled individually. CSS is then used to achieve the same effect
(automatically sizing the element to fit the original pgp element).